move core logics to cli.rs
